requirements.txt
# Core NeMo and ML dependencies
nemo-toolkit>=2.0.0
nemo-run>=0.4.0
torch>=2.0.0
pytorch-lightning>=2.0.0
transformers>=4.20.0
huggingface-hub>=0.15.0

# Data processing
datasets>=2.0.0
tokenizers>=0.13.0
pandas>=1.5.0
numpy>=1.21.0
scipy>=1.9.0

# Japanese text processing
mecab-python3>=1.0.0
fugashi>=1.3.0
ipadic>=1.0.0
uzushio>=1.0.0

# Experiment tracking and visualization
wandb>=0.15.0
tensorboard>=2.10.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Configuration and utilities
hydra-core>=1.2.0
omegaconf>=2.3.0
pydantic>=1.10.0
rich>=13.0.0
typer>=0.9.0
tqdm>=4.64.0
loguru>=0.6.0

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
isort>=5.10.0
pre-commit>=2.20.0
mypy>=0.991

# Documentation
sphinx>=5.0.0
sphinx-rtd-theme>=1.0.0
myst-parser>=0.18.0

# Optional: for advanced features
apex>=0.1  # NVIDIA Apex for mixed precision training
flash-attn>=2.0.0  # Flash Attention for efficient attention
triton>=2.0.0  # Triton for custom kernels

# Development tools
jupyter>=1.0.0
ipykernel>=6.0.0
ipywidgets>=8.0.0
notebook>=6.0.0 